Remove runtime/language tests; will move relevant tests to their respect repo.
