+++ tests/src/test/scala/system/basic/WskConductorTests.scala
@@ -19,31 +19,22 @@ package system.basic
 
 import scala.concurrent.duration.DurationInt
 import scala.language.postfixOps
-
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
-
-import common.ActivationResult
-import common.StreamLogging
-import common.JsHelpers
-import common.TestHelpers
-import common.TestUtils
-import common.WskOperations
-import common.WskProps
-import common.WskTestHelpers
-
+import common._
+import common.rest.WskRestOperations
 import spray.json._
 import spray.json.DefaultJsonProtocol._
-
 import whisk.core.entity.size.SizeInt
 import whisk.core.WhiskConfig
 import whisk.http.Messages._
 
 @RunWith(classOf[JUnitRunner])
-abstract class WskConductorTests extends TestHelpers with WskTestHelpers with JsHelpers with StreamLogging {
+class WskConductorTests extends TestHelpers with WskTestHelpers with JsHelpers with StreamLogging with WskActorSystem {
 
   implicit val wskprops = WskProps()
-  val wsk: WskOperations
+  val wsk: WskOperations = new WskRestOperations
+
   val allowedActionDuration = 120 seconds
 
   val testString = "this is a test"
