+++ tests/src/test/scala/whisk/core/cli/test/WskWebActionsTests.scala
@@ -23,17 +23,11 @@ import java.util.Base64
 import scala.util.Failure
 import scala.util.Try
 import org.junit.runner.RunWith
-import org.scalatest.BeforeAndAfterAll
 import org.scalatest.junit.JUnitRunner
 import com.jayway.restassured.RestAssured
 import com.jayway.restassured.response.Header
-import common.TestHelpers
-import common.TestUtils
-import common.WhiskProperties
-import common.WskOperations
-import common.WskProps
-import common.WskTestHelpers
-import common.SimpleExec
+import common._
+import common.rest.WskRestOperations
 import spray.json._
 import spray.json.DefaultJsonProtocol._
 import system.rest.RestUtil
@@ -45,11 +39,11 @@ import whisk.core.entity.Subject
  * Tests web actions.
  */
 @RunWith(classOf[JUnitRunner])
-abstract class WskWebActionsTests extends TestHelpers with WskTestHelpers with RestUtil with BeforeAndAfterAll {
+class WskWebActionsTests extends TestHelpers with WskTestHelpers with RestUtil with WskActorSystem {
   val MAX_URL_LENGTH = 8192 // 8K matching nginx default
 
   private implicit val wskprops = WskProps()
-  val wsk: WskOperations
+  val wsk: WskOperations = new WskRestOperations
   lazy val namespace = wsk.namespace.whois()
 
   protected val testRoutePath: String = "/api/v1/web"
